LLM_Quantized_Evaluation/requirements.txt
langchain==0.3.19
pytest==8.4.1
deepeval==3.3.2

openai>=1.0.0,<3.0.0

detoxify==0.5.2
Dbias==0.1.5
loguru==0.7.3
coverage==7.10.1

tensorflow
python-dotenv